Changed structure in admin controller and moved styles to separate file
