+++ GiEnJul/Features/RecipientRepository.cs
@@ -2,15 +2,15 @@
 using GiEnJul.Infrastructure;
 using Serilog;
 using System.Threading.Tasks;
-
+using Microsoft.Azure.Cosmos.Table;
 namespace GiEnJul.Features
 {
     public interface IRecipientRepository : IGenericRepository<Entities.Recipient>
     {
         Task<Models.Recipient> DeleteAsync(Models.Recipient model);
         Task<Models.Recipient> InsertOrReplaceAsync(Models.Recipient model);
+        Task<Models.Recipient> GetUnmatchedRecipientsAsync(string location, string currentEvent);
     }
-
     public class RecipientRepository : GenericRepository<Entities.Recipient>, IRecipientRepository
     {
         public RecipientRepository(ISettings settings, IMapper mapper, ILogger log, string tableName = "Recipient") : base(settings, tableName, mapper, log)
@@ -27,5 +27,16 @@ namespace GiEnJul.Features
             var deleted = await DeleteAsync(_mapper.Map<Entities.Recipient>(model));
             return _mapper.Map<Models.Recipient>(deleted);
         }
-    }
-}
+        public async Task<Models.Recipient> GetUnmatchedRecipientsAsync( string location, string currentEvent )
+        {
+            var query = new TableQuery<Entities.Recipient>() { FilterString = 
+                TableQuery.CombineFilters(
+                    TableQuery.GenerateFilterCondition("PartitionKey", QueryComparisons.Equal, $"{currentEvent}_{location}"), 
+                    TableOperators.And, 
+                    TableQuery.GenerateFilterConditionForBool("IsMatched", QueryComparisons.Equal, false))
+            };
+
+            var recipients = await GetAllByQueryAsync(query);
+            return _mapper.Map<Models.Recipient>(recipients);
+        }
+}}
